requirements-dev.txt
-r requirements.txt

# Code style / formatting
black==24.10.0
isort==5.13.2
flake8==7.1.1

#Testes
pytest==8.3.3
pytest-django==4.9.0
mock==5.1.0

# Better Dev Experience
ipython==8.28.0

# CORS (para quando o frontend está em domínio diferente)
djangorestframework==3.15.2
